gadget,overlord: do not try to update if no content is defined
Make sure to avoid additional progress in the moment we know an update
will not happen as no content is defined in a structure, so we do not
need to look for mount points for such structures, etc. This change
forces to make some changes in tests to make them more
realistic (content needs to be provided to reach some code paths while
this was not necessary before this change).