fix store-state tool restoring fakestore
Some tests fail before the fake store is configured so the dir
"STORE_DIR" is nor created. When the test is restored, the test fails to
restore because the tool store-state is returning 1 in case the
STORE_DIR dir does not exist.

This change makes more robust the store-state tool to manage this
scenario and continue in case the fake store was not configured
properly.