snapstate/flags: add a new flag for snap states
ImplicitlyInstalled will be set to true just in case the snap being
installed as a result of being needed by some other snap, as with
content (default) providers, base snaps.

If a user specifically installs a snap, it is to be marked as explicitly
installed, which means ImplicitlyInstalled is false.

Signed-off-by: Joseph Brock <joseph.brock@canonical.com>
Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>