bootloader/assets/data/grub-recovery.cfg: set root when chainloading
Upstream grub does not pass to LoadImage the correct path to the
loaded binary. This results in second grub not knowing where it was
loaded from and loading the wrong grub.cfg. If we set "root", then the
path becomes correct.

Currently in mantic, the Ubuntu patch that used to be applied to 2.06
is not applied to 2.12. So we are we are back to upstream behavior.

The common use case for chainloading in EFI with grub (see
30_os-prober) always sets root. So we should do it and not get broken
again.