cmd/snap-update-ns: put mount entries requiring mimics before independent ones
Addresses https://bugs.launchpad.net/snapd/+bug/2034056 by placing mount
entries that don't require the creation of mimics after the ones that
do. In the bug report, a layout entry that depended on the content entry
was ordered before its dependency because entries that don't need mimics
(the layout entry) were going before.
Although this may fix the most common failure scenarios, it's not 100%
correct because it a layout entry can also require a mimic to be created
in which case it may be sorted before if its target path is
lexicographically smaller than the content's without being a parent of
it. This is documented in the code, in a unit test and in the bug
report. Since it'll require more complex logic, this commit mitigates
the issue for most cases and I'll open a follow-up PR later.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>