o/snapshotstate: Discard namespace after snap restore
The root folder in the snap namespace is bind-mounted to
/var/snap/<-snap->/<-rev->/root. During the 'cleanup-after-
restore' step in snap restore, the source folder is deleted,
causing the root folder's size in the snap namespace to
become zero and preventing the creation of files. To ensure
normal operation after snap restore, discarding the namespace
is essential for snap-confine.

Signed-off-by: Chengen Du <chengen.du@canonical.com>