tests: clear snap cache before interfaces-content test
The test was pulling in the core snap which failed to be validated.
This happened because the snap was already in the cache and had
been modified in the prepare stage. Since it was modified, there's
no assertion matching the digest. Clearing the cache fixes the
problem. This was triggered because we recently stopped building
core on i386 so all channels now have the same revision.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>