gadget: prevent failures when updating kernel/gadget on UC16/18
Make sure that kernel/gadget updates on UC16/18 do not fail even in
the case that the gadget volumes do not match the installed disk. This
fixes a regression in behavior due to DiskTraitsFromDeviceAndValidate
now checking compatibility with the disk, which was happening in a
later stage in the past.