many: add components side info to snapstate
Add components information to the snap state. This is included in the
Sequence variable, so we need to handle in a special way the
marshaling to keep compatibility with state files created by older
snapd versions.

The change to the Sequence field implies multiple changes in tests
that use it. Ancillary methods to handle this have been created.