strutil: add `MultiCommaSeparatedList`
This new function is useful for parsing `url.Values` where there may be
multiple values for the same key, and those values may themselves be
comma-separated values.

This is a port of a commit from the `x-go` package, which was originally
intended to be used in pebble. The commit can be found here:

https://github.com/canonical/x-go/commit/32684ae6fb98b1a3e67313618423a13b30102ce3

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>