snap/pack: fix pack with integrity unit test on non-Ubuntu
Apparently on non-Debian/non-Ubuntu systems, the squashfs image file is padded
to 64k by default. Snap does not pass -nopad to squashfs, hence the created
image is in line with however squuashfs was set up on the host.

This fixes the unit test failure on openSUSE blocking the package update since
2.60.

Signed-off-by: Maciej Borzecki <maciek.borzecki@gmail.com>