aspects: match requests on prefixes of patterns
Previously, requests had to match an aspect's access pattern precisely
(not including placeholders). This relaxes the request matching to first
look for an exact match and then allow a match if the request is a
prefix of the pattern. Currently, only the result will be returned but
once this is changed so the result is returned in a namespace given by
the pattern, the matching request prefix will be used to cut through
this namespace.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>