cmd/snap-update-ns: set sythentic mounts x-snapd.needed-by to entry id instead of path
snap-update-ns was dropping some synthetic mounts due to failure to match
desired mounts that depend on them.

This mismatching was due to setting x-snapd.needed-by to whatever path that needed it
(source or target) and not the mount entry id (XSnapdEntryID() which defaults to target
if not set) which is later on is used in NeededChanges() to compute which synthetic
mounts need to be kept.

partially fixes: https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/2043993

mount ordering/logic needs a more in-depth refactor, this is just a bug fix in the
existing implementation.

Signed-off-by: Zeyad Gouda <zeyad.gouda@canonical.com>