boot: prune some bootchains that should not happen
We are supposed to update on disk from the back of the boot chain to
the beginning. That means and old shim should not be expected to
install a new grub.

We can reduce the list of possible boot chains we give to secboot.
When we will switch to secboot master, secboot will fail if
it sees boot chains that cannot work. So we should make sure not
to pass them.

There is still a bug where we update the seed partition before boot
partition. But here we assume the bug is fixed.