aspects: include paths in validation errors (#13360)
* aspects: include paths in validation errors

Add path prefixes to validation errors to help users identify the
element breaking the aspect's constraints. This turned out to be
a bit more complex than I initially thought because elements aren't
necessarily parsed in a top-down, BFS-like way. This is because there
are user-defined types which are parsed first and then referenced in
other definitions. So these pre-parsed schemas need to be updated with
the correct path when they are referenced and container types (maps and
arrays) need to propagate that down to the parsers that they contain.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* aspects: improve validation error message

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* aspects: include indexed array in validation error path

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* aspects: simplify validation error construction

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* aspects: minor simplifications

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* aspects: minor cleanups

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* aspects: add validation error helpers

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* aspects: export and rename field

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

---------

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>