aspects: Get() matches request on prefixes and returns results in a merged namespace (#13378)
* aspects: Get matches on prefixes and returns namespaced results

Get() can match requests to prefixes of the "request" field specified in
the aspect's rules. When it does, the result is returned in a namespace
equivalent to the unmatched sub-keys. Conceptually, the specified
"request" is a namespace that the user's request cuts through during the
operation. If there is more than one match for a request, the results
are merged when possible (e.g., maps are merged and scalar can replace
other scalars in leafs).

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* aspects: improve var and err messages

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* aspects: improve var name and comments

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* aspects: minor simplifications and improvements

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

---------

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>