o/state,daemon: add per-user notices
This is a port of per-user notices from Pebble.

Relevant pebble PR: https://github.com/canonical/pebble/pull/323

We wish to be able to restrict the visibility of a given notice to
particular users. This PR adds to `Notice` a `userID` field which
corresponds to the UID of the intended recipient of the notice
(usually its creator). The `userID` is stored as a `*uint32`, where
`nil` indicates that the notice is public, and thus accessible to all
users.

The `NoticesFilter` includes a `UserID *uint32` as well, and if it is
not `nil`, then only public notices and notices matching the filter
`UserID` are returned. Public notices are always returned regardless
of the `UserID` filter, so long as they match the other filters as
well.

The root/admin users have access to all notices, while other users may
only view their own notices or public notices. This is enforced by
the API, which by default populates the `UserID` filter with the
request UID. The state is not aware of the permissions of admin or
non-admin users, and treats all non-`nil` UIDs identically.

Instead of their own notices, admin users may get the notices of
another user by passing the `user-id=<uid>` argument to the API at
`/v1/notices`. To get all notices for all users, an admin may pass
`select=all` instead, though they cannot use both `user-id` and
`select` in the same request.

When creating a notice via the API, the request UID is used as the
`userID` of the notice. There is no way to create a public notice via
the API.

The client and CLI correspond closely to the API. The only distinction
is that when using the `pebble notice TYPE KEY` variant, there is now
an optional `--uid UID` argument (admin only) which lets an admin
look up by type and key a particular notice from another user.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>