interfaces: update desktop interface AppArmor permanent slot rules (#13308)
This PR combines a few changes that we have been using on the Core Desktop images for a while. This includes:

Allow snaps with a desktop slot to talk to xdg-desktop-portal. This avoids the need for a self-connected desktop plug on the snap to allow e.g. opening links from apps bundled in the ubuntu-desktop-session snap.

Allow access to GDM's OpenReauthenticationChannel API and the associated private bus address. This is used by gnome-shell to authenticate the user on its lock screen. To work with newer GNOME versions, we depend on GDM !226 to use a socket location outside of /tmp (which can't work with a snap-style private tmp directory).

Allow communication with colord on the D-Bus system bus. The shell uses this to access color profiles associated with the display.

Allow binding to various D-Bus session names used by gnome-shell and gnome-settings-daemon. We had been using dbus interface slots for some of these, but I think it makes sense to grant them via permanent slot rules:
* There is a regular naming scheme for these names, and more are added release to release.
* None of the names support D-Bus activation.
* For the most part, the names are used for communication between different components of the desktop shell. Where it might make sense to allow communication with other application snaps, we'd probably want to handle it through interfaces like desktop rather than through these generic dbus slots.


* interfaces: give snaps with a desktop slot access to xdg-desktop-portal

org.freedesktop.portal.OpenURI doesn’t work in core desktop.
For example, opening Gnome Control Center->Privaty->Location
Services and clicking on “Privacy Policy” doesn’t work.

* interfaces/desktop: allow access to GDM for lock screen support

* interfaces: allow property access to GDM's D-Bus interface

* interfaces: grant access to the OpenReauthenticationChannel method and corresponding socket

* interfaces: allow desktop slot access to new location of GDM dbus sockets

* interfaces: allow desktop slot to bind to various gnome-shell D-Bus names

* interfaces: allow dbus slot to talk to colord on the system bus

* interfaces: allow desktop slot to bind gnome-settings-daemon D-Bus names

* interfaces: updates based on Zeyad's review

---------

Co-authored-by: Sergio Costas Rodriguez <sergio.costas@canonical.com>
Co-authored-by: Ken VanDine <ken.vandine@canonical.com>