many: add support for AppArmor unconfined profile mode (#13333)
* sandbox/apparmor: read apparmor kernel sub-features

These are required to be able to detect whether the running kernel supports the
use of the unconfined profile mode amongst others.

Signed-off-by: Alex Murray <alex.murray@canonical.com>

* sandbox/apparmor: add support for probing parser for unconfined flag

AppArmor 4.0 introduced a new flag named unconfined - since this is a profile
flag rather than part of the main profile syntax, also add support for probing
for flags within AppArmor profiles.

Signed-off-by: Alex Murray <alex.murray@canonical.com>

* interfaces/apparmor: add support for unconfined profile flag

Add a new unconfined attribute to the AppArmor specification which is used to
denote that a profile should include the unconfined flag. Then refactor flag
handling within the AppArmor backend so that both this new flag and the complain
flag (used for classic and dev mode snaps) are handled together. Finally, only
emit this flag when we are sure both the AppArmor parser and kernel provide
appropriate support.

Signed-off-by: Alex Murray <alex.murray@canonical.com>

* interfaces/builtin/lxd_support: add userns and unconfined permissions

LXD requires the use of unprivileged userns and the ability to run unconfined -
so support for both of these to the lxd-support interface.

Signed-off-by: Alex Murray <alex.murray@canonical.com>

* tests/main/lxd: verify apparmor profile has expected perms on mantic

On Ubuntu 23.10 the LXD snap should run with the new unconfined profile flag and
with the userns AppArmor permission, so check for these via spread.

Signed-off-by: Alex Murray <alex.murray@canonical.com>

* interfaces: rework AppArmor unconfined mode to be a static property

As suggested by @pedronis, refactor the implementation of AppArmor unconfined
profile mode to be a static property of an interface so that it is much more
tightly coupled to the interface itself. Add various tests as well to ensure
that only super-privileged interfaces can use this property too.

Signed-off-by: Alex Murray <alex.murray@canonical.com>

* tests/main/lxd: add details to the spread test

Signed-off-by: Alex Murray <alex.murray@canonical.com>

* sandbox/apparmor: test kernel sub-features

Signed-off-by: Alex Murray <alex.murray@canonical.com>

* interfaces: clarify comment regarding AppArmorUnconfined{Plugs,Slots}

Signed-off-by: Alex Murray <alex.murray@canonical.com>

* interfaces/policy: check for unconfined plug/slots outside type loop

Signed-off-by: Alex Murray <alex.murray@canonical.com>

* sandbox/apparmor: fix double space in profiles when probing parser

Thanks for the suggestion @MiguelPires

Signed-off-by: Alex Murray <alex.murray@canonical.com>

---------

Signed-off-by: Alex Murray <alex.murray@canonical.com>