aspects: Set can write to multiple paths
Allow writes to match several entries with the same request but
different paths so the storage can change while supporting old and new
paths. The matching should also work on prefixes so we have to filter
out matching paths that are nested inside others and use unmatched
request prefixes to access the nested part of the value to store.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>