s/apparmor/notify{,/listener}: add support for multiple requests at once
The kernel may wish to send multiple apparmor messages in a single ioctl
buffer. This commit adds handling for this via new methods in the notify
message code. Additionally, this exposes a method to get the length of
the first message in a data slice.

Since the existing concurrency tests use write syscalls on a file
descriptor, this new handling allows the concurrency tests to pass now
as well, even if multiple requests are written to the notify file
descriptor before they are read by the listener again.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>