Group the systems used in the github workflow
The idea of this change is to reduce the number of github runners.

The systems in 1 group will be executed together in the same spread run,
so in this case the number of runners needed to executed the whole
change is reduced from 29 to 21

This approach could be also used to group in a different way the
systems.