s/apparmor/notify/listener: do not send auto-denies when listener closed
The kernel should be responsible for sending back deny responses to
applications if the listener is closed before it sends back a response
to the kernel. Better yet, if a new listener is registered, the kernel
can resend previous unanswered requests to the new listener, so there is
no signal loss.

We still want `Close()` to wait until all tracked goroutines terminate
before closing the notify file, since the listener may have just
received a reply that we might as well finish up passing along to the
kernel. And, that way those tracked goroutines won't be hit with
additional errors writing to a closed file, and `Run()` and `Close()`
won't return until every thread from the listener is closed down.

This greatly simplifies testing and to a lesser extent, the listener
itself. It will yield much greater flexibility in the future, however.

As part of this change, we also revert to requests being sent over
`Reqs()` synchronously, rather than in a goroutine. This was done for a
few reasons:

1. We don't need to read the next request from the kernel if there's not
   a listener ready to receive it --- nothing wrong with just letting
   the next request sit after the kernel writes it.
2. Sending requests synchronously guarantees their order.
3. We don't need to send an auto-deny on listener close (kernel does
   this for us) so no need to always call `waitAndRespondAaClassFile`
   even if the listener is closing before the request is received, and
   we can instead return early.
4. If we read in a request from the kernel, nobody reads it from `Reqs()`
   before it times out in the kernel (which sends auto-deny), and then
   something starts reading from `Reqs()`, that request will still be
   sent if that send happens in a tracked goroutine, even though the
   operation has already been denied by the kernel --- this is
   unavoidable for the first request received from the kernel while
   there's no reader on `Reqs()`, but we can at least avoid it for any
   subsequent requests that the kernel might try to send while there is
   still no reader on `Reqs()`.
5. This reduces epoll syscalls to get data that we may not use, in the
   case that there's no reader on `Reqs()`: furthermore, the way
   `epoll.Wait()` is implemented internally, it sends repeated epoll
   requests with a (rather long) timeout so it can check if the epoll
   instance has been closed and stop retrying.
6. If a new listener is registered, the kernel will try to send requests
   to that listener instead, and in the future the kernel may re-issue
   past requests from listeners which were closed without responding to
   the new listener --- at the moment, though, this is not enabled, so
   we may as well read as few requests as is reasonable from the kernel
   so that they can be issued to future listeners rather than be
   responded to with an auto-deny from the kernel after the original
   listener disconnects.

The one drawback I could see is that if something finally starts reading
from `Reqs()`, the listener has to do all the work to read subsequent
requests from the kernel and create waiting goroutines starting at that
point, when it could have instead been pre-doing the work before the
reads to `Reqs()` begins. But that also would create a bunch of waiting
goroutines running concurrently, which may never receive a reply and
will only close when the listener closes.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

s/apparmor/notify/listener: make concurrency test race-free

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

s/apparmor/notify/listener: fix concurrency test mocked function restore

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>