o/State: Configurable refresh inhibition days number
This patch allows to configure the number of days refresh is
inhibited, between 1 and the current 14 days.

This patch implements proposal SD165. The rationaly behind this
is that 14 days is a reasonable value for a server, but not for
a desktop, where users are, usually, less experienced than
sysadmins, and so it is a good idea to keep updated the snaps
faster.

The proposal is available at https://docs.google.com/document/d/1ZdKwvu-WRyAKLahFD7Tq_gEAwdt38VBu4KU6w31TZjo/edit?usp=sharing