agentnotify: show the snap icon when autorefresh is done
When an application has to be closed by the user to allow to do
and autorefresh, the notification has the application icon. But
when the autorefresh has been completed, the notification has
the snapcraft icon instead.

This MR fixes this. With it, the notification specifying that
the snap has been successfully refreshed and that it can be
relaunched shows the application's icon.

Fix https://warthogs.atlassian.net/browse/UDENG-2030