sandbox/cgroup: fix handling of watched paths, ensure proper cleanup, refactor
When setting up watch for a given path, we have kept a reference to the leaf
node, but the watch was really configured for the parent path. However, the
references to the parent path were not being properly tracked. This resulted in
the watch of the parent path never getting cleaned up. While at it, to keep the
bookkeeping simple, since paths are reference counted, we can keep them in a
hash map instead of a list for easier manipulation and lookup.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>