Register dbus interfaces before the well-known name
The GApplication notifications backend registers the
org.freedesktop.Application DBus interface to allow the agent
to be launched with DBus-Activation. But to avoid race
conditions, it is paramount to register it before than the
well-known name, thus ensuring that the interface is available
when the name appears in the bus.