interfaces/apparmor: rework unconfined mode for dynamic enablement
For an interface to use unconfined mode, it must both declare support for it as
a static property and then enable it by an explicit call. This allows interfaces
to enable this dynamically via an plug/slot attribute or similar as needed (or
if not, it can be enabled the permanent plug/slot callback for the interface
instead).

Signed-off-by: Alex Murray <alex.murray@canonical.com>