spread.yaml: use AUTOPKGTEST_NORMAL_USER instead of hard-coded ubuntu
The name of the unprivileged user account differs from one distribution
to another. Recent versions of autopkgtest use a special "user" user
account and provide an environment variable to locate the correct name.

Retain the former value as default value in case the variable is not
set in the Ubuntu infrastructure.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>