debian: cope with unset ID and VERSION_ID
There is no guarantee that ID and VERSION_ID are set in /etc/os-release.
Provide defaults for ID and VERSION_ID so that spread runs with _some_
system name. Spread rejects system names with two consecutive dashes, so
use a special-case for unset VERSION_ID since we know it is used in a
context of the Debian archive.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

fixup-sid-autopkgtest