cmd/snap: improve stdout/stderr handling when not processing strace output
There are scenarios in which we are not processing the output of strace command,
specifically when the user has passed `--raw`. The user can also redirect strace
output to a file passing `-o` || `--output`, in which case processing is not
possible either. In both scenarios it makes no sense to set up a pipe on
stdout/stderr.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>