o/devicestate, overlord, daemon: add flag to devicestate.Remodel to force an offline remodel (#13494)
This flag will enable offline remodeling, even when all snaps required for the remodel are not provided locally. The remodel will fall back to using already installed snaps in situations that it can. This includes simple channel switches, and when the required revision of a snap is already installed. If a snap is required for the remodel, and the snap is neither provided locally or already installed, then the remodel will fail.

* snap: add support for errors.Is for NotInstalledError

* o/snapstate: add Type to SnapSetup created in snapstate.Switch

* o/devicestate, overlord, daemon: add flag to devicestate.Remodel to force an offline remodel

This flag will enable offline remodeling, even when all snaps required
for the remodel are not provided locally. The remodel will fall back to
using already installed snaps in situations that it can. This includes
simple channel switches, and when the required revision of a snap is
already installed.

* o/devicestate: re-word comment on RemodelOptions.Offline

* o/devicestate: convert sideInfoAndPathFromID to a method on remodelVariant

* o/devicestate: make comment in UpdateWithDeviceContext more clear

* o/devicestate: add TODO about looking for other revisions

* o/devicestate: add comment explaining that case in UpdateWithDeviceContext is unexpected

* o/devicestate: prefix unexpected state error with "internal error"

* o/devicestate: use var for default initialization

* o/devicestate: invert if in UpdateWithDeviceContext to reduce some nesting

* o/devicestate: rename sideInfoAndPathFromID and add some clarifying comments

* o/devicestate: remove unused parameter from remodelEssentialSnapTasks

* o/devicestate: correct copy-pasted comment