daemon,o/ifacestate/apparmorprompting: add interface to requests and rules
Now that the listener includes the names of the interfaces associated with
the rules which triggered the request from the kernel, an interface can
be passed on to users via prompt requests, and can be used to further
differentiate rules.

Generally, interfaces should not provide overlapping rules. If they do,
we still want to identify which interface has priority in the list of
interfaces provided by the listener, and include that interface in the
prompt request and any associated rules.

A request associated with one interface should never be handled by a
rule associated with a different interface. This makes it much easier
for users to reason about which rules apply to a particular request and
to modify or delete rules without worrying about their effects on other
interfaces.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>