sandbox/cgroup: refactor and cleanup, fix path tracking and watch removal (#13508)
This is an bunch of fixes and refactors for the cgroup monitoring code. The main issue addressed is with tracked path removal. We have previously added references to the leaf paths internally, but the actual watch was established using the parent path, for which the reference was never updated. As a result if all leaf paths were gone, we did not clean up the watch on the parent path.

While at it, change the structure internally to use a hash map in the group to watch to simplify bookkeeping.

Next, added support for the context and a way to cleanup and close the watcher in a reasonable manner.

A bunch of improvements to the tests, so that we are no longer guessing and sleeping hoping to observe the right events. Instead provide an observer which the tests can hook to and inspect the state as needed.

* sandbox/cgroup: fix handling of watched paths, ensure proper cleanup, refactor

When setting up watch for a given path, we have kept a reference to the leaf
node, but the watch was really configured for the parent path. However, the
references to the parent path were not being properly tracked. This resulted in
the watch of the parent path never getting cleaned up. While at it, to keep the
bookkeeping simple, since paths are reference counted, we can keep them in a
hash map instead of a list for easier manipulation and lookup.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* sandbox/cgroup: better handling of watcher errors

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* sandbox/cgroup: support context in the watcher

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* sandbox/cgroup: add internal observer to facilitate testing

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* sandbox/cgroup: update tests

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* sandbox/cgroup: debug logs

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* sandbox/cgroup: document error handling quirks

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* sandbox/group: drop a pointless test

Drop a unit test which does not make sense since we cannot confirm that recovery
handling inside the goroutine was successful.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* sandbox/cgroup: improve tests

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* sandbox/cgroup: address review comments

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>